refactor: merge InfiniteOptExts
